requirements.txt
numpy==1.26.4
pandas==2.2.3
matplotlib==3.9.2
Pillow==12.1.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.3
argparse

scikit-learn
scipy

# Notebook support
# jupyter==1.1.1
# ipywidgets==8.1.5

# Optional packages used by specific scripts/notebooks
transformers
seaborn
facenet-pytorch
